Separate class for IBlackPattern to Democles transformation
